+++ manabom/src/main/java/mannabom_server/manabom/domain/meeting/enums/ChatUserStatusConverter.java
@@ -4,14 +4,14 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter
-public class ChatUserStatusConverter implements AttributeConverter<ChatUserStatus,Short> {
+public class ChatUserStatusConverter implements AttributeConverter<ChatUserStatus,Byte> {
     @Override
-    public Short convertToDatabaseColumn(ChatUserStatus attribute) {
-        return attribute==null? null : attribute.getCode();
+    public Byte convertToDatabaseColumn(ChatUserStatus attribute) {
+        return attribute==null? null : (byte) attribute.getCode();
     }
 
     @Override
-    public ChatUserStatus convertToEntityAttribute(Short dbData) {
-        return dbData==null? null: ChatUserStatus.from(dbData);
+    public ChatUserStatus convertToEntityAttribute(Byte dbData) {
+        return dbData==null? null: ChatUserStatus.from(dbData.intValue());
     }
 }
